Run all package tests from project root
Adds a testing package which provides a generic way to run all tests
from the project's root. This is pretty generically useful because we
depend on content, views, layouts, etc. directories that only exist in
one place.

The testing package should be include via:

``` go
_ "github.com/brandur/sorg/testing"
```

(Note the preceding underscore so that naming doesn't clash with the
regular testing package.)