Replace raw Monodraw text with SVGs
Replaces raw Monodraw text with SVGs that were exported from the program
itself.

This is good for two reasons:

* Something in my change in Markdown processors on the way to static
  broke the way Monodraw diagrams get shown. Every one currently in use
  doesn't look right.

* The SVG diagrams look better. Lines are better connected and the
  entire effect is quite a bit cleaner than with the text.

I use SVG instead of PNGs here just because we only need one file
(instead of one for retina and non-retina), and they'll scale better as
we getter higher resolution displays.


Former-commit-id: 2235f31dec81d833be1db474d70640517a3fab12