New flexbox-based layout
Introduces a new layout with a fixed navigation bar on the left, on room
for the top-level tables of contents that are needed by various pages.
The major motivation here is just that navigation is needlessly
difficult right now, requiring users to scroll all the way to the top or
bottom of the page.

I'm still not totally happy with some pages (the home page for example),
and overall the design could still be a lot better. One known problem is
that if the fixed div's content becomes too long and the window is too
short, it's not scrollable.