Switch to dep for dependency management
Switches the project over to the dep tool for dependency management
because it seems likely to eventually be the official way forward:

https://github.com/golang/dep

I could generally not get govendor to do what I wanted to anyway, so
it'll be nice to try something new.

All dependencies have been updated. It seems that I got lucky for most
of them -- blackfriday seems to be the only package that broke its
interface.


Former-commit-id: 39e5fa228fdade639517f6738d7863458f416bb7