Replace all "&mdash;" in Markdown with "--"
The Markdown renderer is escaping the "&" in "&mdash;", so these are all
showing up as literally "&mdash;" when viewing the HTML page, which is
pretty undesirable.

Here we just replace all instances with "--". These are translated to a
real emdash when rendered, and they're easier to read in the source.

Fixes #200.


Former-commit-id: 95cf743c6ffb7de1ea8238db0fd088df3eed7edd