Don't skip download when local target exists
This was done as an optimization, but may cause problems if a download
was only partially completely and left in a bad state.

The caller already optimizes downloads by looking for marker files, so
it shouldn't be that big of a deal to just re-download any original in
cases no such markers exists (it should only happen a maximum of once
per environment).