Use public URL instead of AWS creds to get MozJPEG
The trouble with using S3 creds (which would be preferable because they
allow private access) is that they're not available from PRs that are
not made by me, so the build will break for any third party
contributors.

This way has the downside of making access public, but allows the build
to succeed without the need for AWS credentials.