Kill vertical padding + close button
Kill vertical padding on lightbox images so that it's easier to click
into the "null space" to close them (currently, very annoying that it
seemingly does nothing).

Unfortunately, this also means that the close button has nowhere good to
be. Hide it as well because it then becomes relatively obvious and easy
that lightboxes should be closed by clicking anywhere else.