Pick between see read
Love the article and it's premise.

I noticed a small error and the make a pull-request so as a thank you I opted to contribute back

There was a sentenced that had a small mistake. It contained the combination "see read", which seems wrong to me. I opted to change it to "see", but that is a choice best left to you.

Thanks again for the time to publish your thoughts.