Atoms: Use an alternative lexicographic base32 encoding set for slugs
Here, use an alternative base32 encoding character set for generating
atom slugs that ensures that output slugs sort in the same order as
their source timestamps. The encoding set is the same as RFC 4648 except
that numbers are moved from the end of the set to the beginning.

It's not great the greatest thing in the world to change the existing
slugs at this point, but the project's still new enough that if it's
ever going to be done it should be done right now. Preexisting atoms
will continue to exist in S3 so that outstanding links aren't broken.