Add another derived work for Base32 slugs
This article came to mind recently when I had a need for behavior like
this so I went ahead and wrote a Ruby library that implements it. I used
the same name as Luciano (with permission!) because I liked it quite
a bit and feel that "naming the concept" is a good thing.