Switch back to self-built ImageMagick + HEIC support
Worse than using the official release maybe, but comes with HEIC
support, and possibly a little faster since it's downloaded from GitHub
Actions.

Previously I'd been using a self-built binary, but started getting this
error:

    Run /home/runner/imagemagick/bin/magick identify -list format
    /home/runner/imagemagick/bin/magick: symbol lookup error: /home/runner/imagemagick/bin/magick: undefined symbol: heif_deinit

It appears that the cause was that I'd slipped to using a version of
libheif that was too old, and ImageMagick just has godawful error
messages. Since writing imagemagick-builder, it appears that the
repositories for libheif-dev and libde265-dev have been mainlined,
allowing me to decouple from the 3rd party repositories that I'd been
using and which had become outdated, and go to better maintained,
built-in versions [1]. Doing so fixed the problem above.

This patch also brings back support for HEIC.

[1] https://github.com/brandur/imagemagick-builder/commit/895a6a1b86b0144280b5fe348932b935f74a51b7