Hide "uses" and "now" at small screen sizes
Unfortunately nav is back up to occupying too much of the total width on
a small screen like an iPhone. Since "uses" and "now" are kind of "fun"
pages that probably don't get a lot of clickthrough anyway, hide them at
small screen sizes so the nav appears more centered and balanced.