Fix photo overflowing on about/now/uses pages
Fix a problem where at smaller screen sizes, photos on the about, now,
and uses pages would overflow and create a horizontal scroll bar.

This improves the photo styling, but this could still stand to be
improved in that at small sizes the photos won't expand to the full size
of the frame because they're contained in a div with padding.