Move main layout to the top level of `layouts/`
The main layout was originally conceived for particular pages like the
Belize series and therefore started life in `layouts/pages/`. It's since
become much more general and serves as the base layout for every page on
the new system. Give it an appropriate new home at the top level of the
`layouts/` directory.