Set the theme once more on DOM content load to set toggle
There was a bug in the theme toggle in that it would always to be set to
"auto" on page load, regardless of current state, which prevented
switching to "auto" when starting on either light/dark mode without
first switching to another state.

It turns out that the bug is caused because when we first set theme, the
toggle isn't loaded yet, and therefore not set to the right state. Here,
resolve the problem by setting theme _again_ on DOM content loaded,
which is _a_ solution, even if maybe not the world's greatest one.